Add buttons and style them.
Add the 'Click' and 'Reset' buttons to the web app.
Add functions to be execute when clicking the buttons (just console log print for now).
Create a 'Button.css' file to style the buttons.
Execute 'npm install --save-dev @babel/plugin-proposal-private-property-in-object --legacy-peer-deps' in order to don't see the 'One of your dependencies, babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package ...' warning anymore.