Add element to show number of clicks.
Create a 'Counter' component which shows the value of the 'numberOfClicks' prop received.
Create a 'Counter.css' file to style the number which is showed.
Add a 'Counter' element in 'App.js' to show the number of clicks (for now, it shows a static number).
Add style to the 'click-counter-wrapper' class, to style the number and the buttons of the counter.